+++ tests/commands/export/test_export_causatives_verified_cmd.py
@@ -39,8 +39,10 @@ def test_export_causatives(mock_app, case_obj):
     # variant should NOT be returned
     assert str(variant_obj["position"]) not in result.output
 
-    # Test the cli by providing the right collaborator
-    result = runner.invoke(cli, ["export", "variants", "-c", case_obj["owner"]])
+    # Test the cli by providing the right collaborator, build and variant type
+    result = runner.invoke(
+        cli, ["export", "variants", "-c", case_obj["owner"], "--build", "37", "--category", "snv"]
+    )
     assert result.exit_code == 0
     assert "#CHROM	POS	ID	REF	ALT	QUAL	FILTER	INFO" in result.output
     # variant should be returned