+++ CHANGELOG.md
@@ -8,6 +8,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 ### Added
 - A documentation page for institute settings (#6212)
 - Parse and display Paraphrase output from Nallo Paraphase conversion (#6178)
+- Command line option to export causative variants by category and/or genome build (#6202)
 - ClinVar germline submissions allows assigning 3 new VUS terms: VUS-high, VUS-mid and VUS-low (#5994)
 ### Changed
 - Genome build is now shown on variant verification "Sanger" emails (#6194)